Improve css positioning in popup (#489)
* Render popup sections in reliable order

Previously, the render order was nondeterministic, because it was rendering on a first-come-first-served basis and the underlying logic had network and other async calls. Now, they all render inside pre-defined containers, so that no matter when the content becomes available, their order stays the same.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>

* Use normal-flow positioning, rather than absolute

Because we now know the order of the components, we no longer need to position them absolutely. Now we can use their standard flow behaviour, and no longer need to enforce fixed heights to parent components depending on their contents.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>

* Improve items vertical sizing in the popup

Plus minor improvements for future updates.

Signed-off-by: Emanuele Feliziani <feliziani.emanuele@gmail.com>