Referrer Truncation (#511)
Add referrer trimming functionality to reduce tracking as users browse the web. 

This feature has two components: 
 * Before headers are sent with any request, determine if the referrer header should be trimmed or not. Trimming generally removes all path and parameter information from the referrer
 * During a page load, modify the JS API document.referrer to the trimmed referrer.

Referrer is not trimmed when the request or page load is first party (same domain or owner). If the request or page is to a known tracker, the referrer is trimmed to only eTLD+1 (excludes subdomains)

Co-authored-by: Konrad Dzwinel <konrad@duckduckgo.com>