Reduce ways page styles can influence our CTP placeholders
The Click to Play placeholder elements are injected into the page, so
the website's own stylesheets can change how they look. Something as
simple as

    div { border: 2px red solid; }

would be quite problematic. With these changes, we put as much of our
placeholder interface inside a shadowRoot as possible. The website's
styles shouldn't apply to anything in our shadowRoot. We also take
care to reduce the number of wrapping div elements, and to ensure the
remaining exposed divs are at least given the default style of no
border/padding/margin.