Name Click to Play events consistently
It's better that we use a consistent prefix for the Click to Play
related events. That will make it less likely event names will clash
accidentally and it will make naming easier as we add more events. It
also makes their use easier to find in the codebase.

Let's go with "ddg-ctp-" and "ddg-ctp-ENTITY-" as the prefix for all
Click to Play events.