Avoid triggering our Click to Play load event listener multiple times
At the point that the user clicks a Click to Play placeholder, we
reinsert the previously blocked content into the page and wait for it
to load. It turns out our load event listener can fire multiple times
as a side-effect of how we reinsert the content. That is not
intentional and we should ensure that our load event listener only
fires once.