Click to Play code improvements 2 (#954)
* Name Click to Play events consistently

It's better that we use a consistent prefix for the Click to Play
related events. That will make it less likely event names will clash
accidentally and it will make naming easier as we add more events. It
also makes their use easier to find in the codebase.

Let's go with "ddg-ctp-" and "ddg-ctp-ENTITY-" as the prefix for all
Click to Play events.

* Avoid triggering our Click to Play load event listener multiple times

At the point that the user clicks a Click to Play placeholder, we
reinsert the previously blocked content into the page and wait for it
to load. It turns out our load event listener can fire multiple times
as a side-effect of how we reinsert the content. That is not
intentional and we should ensure that our load event listener only
fires once.