Split apart createReplacementWidget function and add events
Split out the DuckWidget creation and element replacing logic from
createReplacementWidget function. Tidy up some of the function and
variable names to make the code easier to understand. Begin emitting
events when a tracking element is replaced, so that the surrogate
scripts can tie tracking elements to their corresponding placeholder.