Add API schema tests for Click to Load surrogate scripts (#1041)
The Click to Load feature redirects requests to certain scripts in
order to replace them with our "surrogate" scripts. It's important we
know when the original script has changed materially, so that we can
update the corresponding surrogate script. To that end, let's add
integration tests that check the schema of important Objects provided
by those scripts against the schema that we expect.