Move integration test configurations into separate JSON files (#1055)
Some integration tests depend on parts of the extension
configuration. We set those parts of the configuration manually, to
ensure the tests will run consistently. In the past, the test
configuration was hard-coded in the integration test files, but here
we move the test configurations out into separate JSON files for
improved readability.