Add Tracker Blocking support
Add Tracker Blocking support, by converting a blocklist[1] into a
declarativeNetRequest ruleset that blocks/allows requests consistently
with the blocking algorithm[2] and the MV2 extension's
implementation[3].

Since this was tricky to get right in practice, we ran the generated
declarativeNetRequest ruleset against test data with ~970k requests to
catch all edge-cases. To facilitate that, I created a
declarativeNetRequest match outcome testing API in Chromium[4], which
I'm in the process of upstreaming.

In order to keep the ruleset size down, I included several
optimisations including:
 - Skipping redundant declarativeNetRequest rules.
 - Combining similar declarativeNetRequest rules where possible
 - Omitting rule conditions where possible

That said, there are some areas that could be improve in the
future. In particular, it would be nice to:
- Better handle tracker entry rules that are anchored to a subdomain
  of the tracker entry's domain.
- Skip excluding initiator domains that are the subdomain of already
  excluded initiator domains for the declarativeNetRequest rule.

1 - E.g. https://staticcdn.duckduckgo.com/trackerblocking/v2.1/tds.json
2 - https://github.com/duckduckgo/duckduckgo-privacy-extension/blob/develop/docs/blocking-algorithm.md
3 - https://github.com/duckduckgo/privacy-grade/blob/4d28937/src/classes/trackers.js
4 - https://crbug.com/1312970