Add Tracker Blocking Allowlist support
The extension configuration[1] allows Tracker Blocking to be quickly
disabled for websites where breakage is reported. Let's add support
for that here, by converting an extension configuration (specially the
trackerAllowlist section) into a declarativeNetRequest that disables
Tracker Blocking where required, taking care to be consistent with the
MV2 implementation.

Notes:
 - We did not have test data for the Tracker Blocking Allowlist, so it
   is possible that some subtle edge-cases have been missed. If
   possible, we should run this against some test data in the future.
 - URL filters could be further improved so that '*' matches literally
   and so that case-insensitive matching isn't used unless necessary.

1 - https://staticcdn.duckduckgo.com/trackerblocking/config/v1/extension-config.json
2 - https://github.com/duckduckgo/duckduckgo-privacy-extension/blob/0438991/shared/js/background/allowlisted-trackers.es6.js