Add session storage wrapper (#1165)
With Chrome Manifest V3, the extension's background page is replaced
with a ServiceWorker. That ServiceWorker is destroyed and recreated
regularly, so the extension's state can no longer be maintained in
memory in the same way. Instead, the new chrome.storage.session[1] API
must be used to keep track of any state that can't be preserved to
disk.

Since chrome.storage.session is not available on other platforms yet,
let's add wrappers that fall back to a Map. That way, the extension's
state can be stored in a consistent way for all platforms.

This change is heavily based on @jonathanKingston's earlier work[2]. Full
credit to Jonathan for this change.

Note: The storage.session API is not Promisified by webextension-polyfill
      like may of the other wrapped APIs. That is OK however, since
      storage.session (and many other extension APIs) support Promises
      natively with Chrome Manifest V3.

1 - https://developer.chrome.com/docs/extensions/reference/storage/#property-session
2 - https://github.com/duckduckgo/duckduckgo-privacy-extension/pull/863