Use `globalThis` instead of `window` in integration tests
With Chrome Manifest V3, the extension's background page is replaced
with a background ServiceWorker that no longer has access to a
`window` Object. Let's replace (where possible) references to the
`window` Object (and implicit references to `window.dbg`) with
references to `globalThis`. While this isn't enough to get the
integration tests passing for Chrome MV3, it removes one more reason
for the tests to fail!

Note: Some references to `window` would have worked, since they are in
      a page instead of the background ServiceWorker. `self` would
      have worked too. Let's replace those anyway, to set a clear
      example about which global Object to use.