Use backgroundWait and pageWait integration test helpers consistently (#1231)
Use the backgroundWait and pageWait integration test helpers
consistently. That way, we can avoid repeating code and avoid failing
when pages/requests load too slowly. It also means that we're properly
waiting for the extension configuration to have loaded before running
the tests.

Along with that, expand the pageWait test helpers to provide more
useful output on timeout and to provide a waitForNetworkIdle wrapper.