Add createAlarm wrapper to avoid overwriting existing alarms (#1241)
With Manifest v3, the background ServiceWorker restarts often. It
turns out that the code that sets up periodic alarms is re-run each
time the ServiceWorker restarts, and that causes the alarms to
reset - they are always X minutes from firing, but will never fire!

We need to take care to check an alarm doesn't already exist, before
attempting to create it. Let's add a createAlarms wrapper to make that
easy to get right.

See also this great writeup by wOxxOm:
https://stackoverflow.com/questions/66391018/how-do-i-call-a-function-periodically-in-a-manifest-v3-chrome-extension/66391601#66391601