Include update timing information in TDS + HTTPS storage classes
The TDS and HTTPS storage classes handle both storing and periodically
updating the extension configuration files. In order to know when an
update is required, the first step is to know both when a
configuration file was last updated and how frequently the
configuration file should be updated.

For that, let's:
- Move the update period into the respective classes.
- Modify loadExtensionFile to expose the Date response header.
- Take note of the lastUpdate time each time a configuration is
  fetched (even if there weren't any changes).

Note: It is suboptimal how we are saving the lastUpdate field in
      localStorage via the settings abstraction, instead of in
      IndexedDB along with the rest of the configuration data. We're
      doing it this way for two reasons.
      1. Firstly, we need to update the lastUpdate time even if there
         was no changes to the configuration. It seems wasteful to
         rewrite the entire configuration, when it hasn't changed,
         just to update the lastUpdate time.
      2. The entity tags for configurations are already stored in this
         way anyway, so it seems pragmatic to do the same with these
         changes.