Make use of locally saved TDS and HTTPS storage on extension startup (#1234)
When the extension starts up, `httpsStorage.getLists()` and
`tdsStorage.getLists()` are called to load some of the extension's
configuration files. Those files are necessarily fetched remotely when
the extension is first installed, and periodically to keep them up to
date. But until now, the files were also fetched each time the
extension was restarted.

That wasn't a problem before, since with MV2 extensions the background
page can persist for a long time. But with the transition to MV3 that
is no longer the case - the background ServiceWorker is restarted
constantly. So let's tweak the `getLists()` methods to accept a
`preferLocal` parameter to avoid the files being fetched every time
the background ServiceWorker is restarted.

Note: Since entity tags were being utilised, the responses were 304
      Not Modified and didn't use much bandwidth. But nevertheless, we
      want to minimise needless requests.