Add comment about session storage for tracking parameter protection
The tracking parameter protection feature processes the list of known
tracking parameters when the are updated and stores them in more
efficient data structures ready for use. They are so far stored in
local variables, which is not MV3 compatible[1].

Luckily however, the code is not required for the MV3 extension, since
tracking parameter protection will need to use the
declarativeNetRequest API there. Let's just add a comment to mention
this, to avoid setting a bad example.

1 - https://developer.chrome.com/docs/extensions/mv3/migrating_to_service_workers/#state