Fix the Facebook SDK schema test
The Facebook Click to Load test page has recently been updated to fix
much of the broken content. That has had the knock on effect of
meaning the page takes longer to finish loading. When testing the
Facebook SDK schema, let's not wait for all requests to finish,
otherwise the chances are it will timeout.