Put email alias fetch attempt count in session storage (#1280)
When the user requests a new @duck.com email address alias, we attempt
to fetch them one. If that should fail, we will retry a few times
before giving up. So far, we kept track of the current fetch attempt
in a local variable, but for Chrome MV3 compatibility we need to move
that to session storage instead.