Properly check the request initiator for tracking parameters protection
We strip known tracking parameters for navigations. Sometimes it is
necessary to disable that if breakage occurs, or if the user has
chosen to disable protections for a website. Both the request URL and
the initiator URL are checked, to see if the feature is disabled, but
a typo prevented the initiator URL from being checked correctly.

For Chrome, the request initiator is in the `initiator`
property[1] (not `initiatorUrl`) and for Firefox it is in the
`originUrl` property[2]. Let's correct that here.

1 - https://developer.chrome.com/docs/extensions/reference/webRequest/#event-onBeforeRequest
2 - https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/webRequest/onBeforeRequest#details_2