Get tests passing on macOS
It turned out that Puppeteer took longer to start Chromium on macOS
than it did on Linux. That was causing the tests to pass out with
Mocha's default two second timeout.

Properly wait for Puppeteer to finish in the before all hook, increase
the timeout there to 20 seconds and get the GitHub test runner to
ensure the tests pass on macOS too.