Fix the settings button in the popup UI for Chrome users
There was a typo in the UI code, which caused the settings button not
to work for Chrome users. The openOptionsPage function took a
parameter named `browser` which contained a string name of the browser
(e.g. "chrome"), but then later attempted to use the
`browser.runtime.openOptionsPage()` API.

Credit to @GioSensation for finding this bug and proposing an initial
fix.