Fix assert.rejects / assert.doesNotReject test cases
Since assert.rejects and assert.doesNotReject are asynchronous
operations, we have to wait for them to complete if the tests are to
complete properly. In correcting that mistake, it turned out that some
of those test cases were not running correctly. With this change all
of those test cases are now running properly and passing.