Inform website that Facebook login failed when user rejects it (#1369)
We warn the user about the privacy implications before they click to
login a website via Facebook. The modal has a "Go back" button, but so
far the website wasn't informed that the login process was
aborted. With this change, we're performing the proper callback and
since the `authResponse` property isn't provided, the website knows
the login wasn't completed[1].

1 - https://developers.facebook.com/docs/reference/javascript/FB.login/v14.0