Reduce the number of files ignored by tsc linting (#1452)
The TypeScript linter tsc is set to ignore some modules, since there
are some linting errors in those files. The problem with that (other
than missing out on the benefits of type checking), is that when one
of the ignored modules is later imported by a non-ignored module, the
previously-ignored module will start being checked again. This results
in a kind of trap where importing the wrong module suddenly results in
dozens of linting errors!

Let's improve the situation by reducing the number of ignored
modules.

One interesting linting error I hit was:

    Module '"X"' declares 'Y' locally, but it is not exported.

It turned out, that was caused by modules that used CommonJS exports,
but that started with an ES import. To fix those errors, I switched
back to CommonJS imports in those situations. We should obviously
switch to ES modules generally in the future.