Fix circular dependencies resulting from importing startup.ready()
The startup module exported a ready() function that could be used to
check if the extension's startup sequence had completed. That was
useful to help prevent race conditions at startup. The problem with
that though, was that the startup module itself imports quite a few
other modules. That resulted in a trap, importing startup.ready() from
the wrong module (one that startup itself needs to import) would
result in a circular dependency and cause peculiar timeouts and
failures.

Split out startup.ready() into the ready module (ready.startupReady()
and ready.startupReadyResolve()). That module can be safely imported
as it has no dependencies of its own. In the future, the ready module
could be also used for other similar ready Promises.