Get request blocking integration tests passing (#1533)
The number of expected blocked requests was hard-coded in the request
blocking integration tests, which meant that they started failing when
the page was updated to make more requests. Let's make use of the test
count that we already have instead, that gets them passing again +
will hopefully avoid the same thing happening next time.