Expand generateDNRRule for GPC protection rules (#40)
The generateDNRRule rule was originally just for generating the
declarativeNetRequest rules necessary for Tracker Blocking. As we
expand to include more privacy features, more rule actions and
conditions need to be supported. Jason is working on adding GPC[1]
protection now, but for that he needs to omit the requestDomains
condition and to specify header modifying actions.

1 - https://help.duckduckgo.com/duckduckgo-help-pages/privacy/gpc/