Rethink some relative rule priorities and expand the tests (#51)
We recently realised[1] that some relative rule priorities weren't
correct. For example, the contentBlocking allowlist was not taking
priority over the tracker blocking allowlist. While discussing those
relative priorities, we agreed it would be nice to have a better
overview of them all too.

So let's adjust some of the rule priorities here, and also take care
to expand the rule priority tests to provide more of an overview.

1 - https://github.com/duckduckgo/ddg2dnr/pull/49#discussion_r1035658292