Avoid throwing exceptions when new tabs are created for MV3 (#1596)
We listen for chrome.tabs.onCreated events, to keep track of which
tabs are open on which URLs. We also listen for
chrome.webRequest.onBeforeRequest events to enforce tracker blocking
and other features on MV2. The onBeforeRequest code-path is also
active for Chrome MV3 builds for now (but non-blocking), though the
aim is to disable that in the future.

In the mean time, it turns out that while on MV2 the tabs.onCreated
event fired reliably before the onBeforeRequest event for a new tab,
with Chrome MV3 that is no longer the case. Let's take care to return
if the current tab isn't known yet, before trying to access the tab's
details.