Fix user allowlist for after upgrade from MV2 to MV3 extension
When the user "allowlists" (disables protections for) a domain, we add
the domain to our user allowlisting declarativeNetRequest rule. That
works fine, but unfortunately does not consider domains that the user
allowlisted before migrating from the Chrome MV2 to Chrome MV3
extension.

We previously handled this case, since we wrongly believed dynamic
rules were lost on an extension update. We since removed the code
however[1]. Let's reapply those changes, to handle the MV2 -> MV3
upgrade case.

1 - https://github.com/duckduckgo/duckduckgo-privacy-extension/commit/8b95cd144051da4d41ae4c2fc4ed6c1f94c88915