Add a note about using a rule ID filter for getDynamicRules
With Chrome 111 there will be a way to filter fetched dynamic/session
rules by rule ID. We should make use of that for
`findExistingDynamicRule()`, since we can query for the desired rule
directly and avoid fetching all the other rules.

Since we currently support back to Chrome 102 for the MV3 extension we
can't use it yet, but let's add a note so we can do that later.