Start using the v4 block list again
Now that the rule matching code in privacy-grade has been updated to
properly ignore unsupported rule actions, it should be safe to switch
back to the v4 block list again. That is necessary to get Click to
Load working for MV3 builds again, and is the next step towards
removing the legacy social_ctp_configuration.json[1] configuration
file.

Note: We also needed to fix the Click to Load YouTube test
      configuration, that was missing the surrogate script entry. It
      turned out that this was working by accident using the v4 block
      list actions in the test block list file.

1 - https://staticcdn.duckduckgo.com/useragents/social_ctp_configuration.json