Refactor the Click to Load declarativeNetRequest rule logic (#1677)
The "Click to Load" feature blocks some embedded third-party content,
while giving users the option to click to load it again. To get that
working with Chrome MV3, we have the necessary allowing
declarativeNetRequest rules ready and add those for tabs where the
blocking shouldn't happen. There are quite a few things to consider,
like if the content is first-party, if the user has allowlisted the
website, if the feature is enabled for the current domain etc. Let's
refactor the declarativeNetRequest rule generation logic to handle
more of those edge-cases.