Fix Chrome build targets after make clean
The manifest.json target in the Makefile copied the locales/ directory
for the chrome target. That clashed with the $(BUILD_DIR)/_locales
target, causing the first chrome build to fail after a `make clean`.

Another issue, was that for chrome-mv3 target some of the required
directories weren't ready in time for the first build after a `make
clean`.

These targets probably need to be tidied up, but as a workaround let's
ensure that only the browsers/BROWSER/*.json files are copied for the
manifest.json target, and that the prerequisite directories are
created for the other problematic targets.