Tidy up the Makefile and ensure updated files are copied again
The Makefile had an issue whereby copied files were not re-copied when
they were updated. This could cause builds to contain out of date
files.

While I was fixing that, I decided to refactor/tidy the Makefile:
- Grouped all top-level targets together at the start of the Makefile.
- Added some missing comments.
- Used whitespace more consistently.
- Moved some out of place targets into the correct section.
- Added missing .PHONY target dependencies.

There is more work to do, in particular it would be nice to use
automatic dependency generation (e.g. `browserify --list`) in order to
reduce regeneration of the bundles.