Stop passing --delete to rsync in Makefile and more tidy up (#1803)
For builds, the Makefile first copies a bunch of files and then builds
some more. For the file copying we used rsync, with the --delete
option. That ended up being problematic, since sometimes the built
files could be deleted by rsync.

While we're at it, let's add/fix some comments and add a few missing
.PHONY dependencies.