Revert recent change to clean target in Makefile (#1832)
There was a bug in the Makefile, whereby `make clean` would run
`rm -rf $(BUILD_DIR)` which usually resulted in `rm -rf build//` being
run. By happenstance this worked, clearing the build directory. I
fixed that, by changing the script to run `rm -rf build` directly.

However, it turns out that the release target also runs `make clean`,
but with the BUILD_DIR variable set correctly. It expected only the
targetted build to be cleared, so when the whole build directory was
cleared instead the release process failed.

For now, let's revert that change, to unblock today's release. In the
future, we should handle this properly.