Avoid triggering Chrome MV3 builds twice
For Chrome MV3 builds, we need to fetch a list of domains that are
then used to generate a "Smarter Encryption" (HTTP -> HTTPS upgrade)
declarativeNetRequest ruleset. We were writing that list of domains as
a text file in the source tree, but that was confusing and could cause
a second incremental build to be triggered during development.

Let's save that text file in the build directory instead, that way
it's clearer, we avoid triggering builds twice and we can remove
the rsync exception.