Ensure surrogates.txt target is triggered each time files are copied (#1839)
We have a problem that sometimes surrogates.txt will be generated
early, before the surrogate scripts are copied into the build
directory. That causes surrogates.txt to be empty/invalid, and
also causes a second incremental build to run during
development. Let's just have the surrogates.txt target depend on
LAST_COPY instead. The cost of regenerating the file unnecessarily for
incremental builds seems worth generating it reliably.

Ideally, we would have targets for all copied files in the
Makefile, and then we could set proper dependencies for the
surrogates.txt target to ensure it's only regenerated when the
surrogate scripts have changed. In practice though, we have a
LAST_COPY target that represents all the file copying.