Ensure contentBlocking "feature" is properly disabled for MV3 builds (#1890)
The contentBlocking "feature" in the configuration is primarily used
to disable tracker blocking (and a few related features) for domains
where users report breakage. During debugging/development however,
sometimes the feature is disabled entirely to disable those features
for all websites. Let's take care to generate the correct allowing
rule for when the contentBlocking feature is disabled entirely.