Generate static MV3 rulesets for tracker blocking
So far, we have managed to keep the necessary declarativeNetRequest
rules for tracker blocking and related features to below the 5,000
dynamic rule limit[1]. However, as we expand the block list and our
protections, the limit is approaching. At the time of writing we're up
to around 4,000 dynamic rules, and that is trending upwards.

Furthermore, for our Safari MV3 port, we will need even more rules for
our protections, since rule conditions such as requestDomains and
excludedRequestDomains are not yet supported.

Therefore, we're going to move to static declarativeNetRequest
rulesets for tracker blocking. As a first step, let's set up the
instrumentation to fetch and update the current/fallback[2] blocklists,
and to generate the corresponding static rulesets.

1 - https://developer.chrome.com/docs/extensions/reference/declarativeNetRequest/#property-MAX_NUMBER_OF_DYNAMIC_AND_SESSION_RULES
2 - During the technical design phase, we decided to add a fallback
    blocklist to MV3 builds as well as the current blocklist. That
    way, if there is a catastrophic problem, we should have a way to
    revert to an older and hopefully less broken blocklist.