Update the uninstall URL each time the ATB setting is updated (#2156)
When a user uninstalls the extension, an survey is displayed which
aims to find out what the user's reason was. Some parameters are
included with that request (e.g. the browser name and version) to
help us diagnose issues. The ATB setting is included as well, so it's
important that we update the uninstall URL each time the ATB setting
is changed. So far, changes to the ATB setting were only reflected
after the next time the updateUninstallURL event fired, which led to a
delay of up to 10 minutes.