Force set list values to false on Site initialization
When recreating a Site object with an existing TabState object, we were not resetting allow- and deny-list values when the URL changed.
This means that the an allowlisted site would have that state persisted to other sites after a navigation.

We can fix this by force setting the value to false when creating the Site object.