Avoid including old smarter_encryption.txt file in builds silently (#2512)
In the past, we wrote shared/data/smarter_encryption.txt when building
the extension and added it to the .gitignore to ensure it wasn't
accidentally included in the repository. Unfortunately however, this
means that if the file exists, it will be somewhat hidden yet be
included in builds. This is a problem given the file's >5MB
size. Let's remove the path from .gitignore, so that the file will be
noticed by developers and removed.