Update @playwright/test dependency to 1.46.0
Unfortunately it seems that we no longer get the details of
chrome-extension://* requests from Playwright. Even when for example
website scripts are redirected to shim (aka "surrogate") scripts
provided by the extension. Previously those requests included
allowed/blocked/redirected status and URLs, but now they seem to
always show up as being blocked and with the URL of
chrome-extension://invalid/. The Facebook Click to Load integration
tests need to be simplified therefore, to just check for allowed
vs blocked requests.