Update privacy-dashboard dependency to 6.0.0
With this update, the way privacy-dashboard communicates with the extension has
changed. Instead of sending messages using the chrome.runtime.sendMessage API,
the chrome.runtime.connect API is used instead. In the future, that will allow
us to listen for the disconnect event as a way to know when the popup UI has
been closed by the browser (e.g. the user clicked away).