Add support for the "toggle reports" feature
When the user clicks to disable protections for a website, the "toggle reports"
feature prompts them to also send a website breakage report to us. This helps
increase the number of user reports we receive for website breakage issues, in
turn helping us fix or mitigate them. The feature also takes care to not prompt
the user too often, at the time of writing only prompting one to three
times (depending on the user's response) every 48 hours.