Bump the minimum supported Chrome version to 128
A bug in the chrome.declarativeNetRequest API[1] meant that extensions could
block requests initiated by other extensions. This cross-extension request
blocking wasn't supposed to be possible[2], and led to lots of issues for users
with multiple browser extensions installed (for example, see this Privacy Badger
bug report[3]).

These cross-extension requests weren't associated with a tabId, and so could not
be differentiated from other such requests - including ServiceWorker initiated
requests. As a workaround, to avoid breaking the other extensions our users
might have installed, we stopped blocking any requests with no associated tabId.

Since then, we have fixed the bug in Chrome[4], and that fix was release with
Chrome 128. So let's increase the minimum supported Chrome version to 128, so
that we can re-enable ServiceWorker initiated request blocking in the future.

1 - https://crbug.com/40896400
2 - https://github.com/w3c/webextensions/issues/369
3 - https://github.com/EFForg/privacybadger/issues/2968
4 - https://chromiumdash.appspot.com/commit/486d638e6977dad73bd207dc914df9319afac152