Update privacy-dashboard dependency to 7.4.0 (#2838)
With this dependency update, the "GetToggleReportOptions" privacy-dashboard
message now also handles the "locale" param ID, so let's start passing that too.

Also, it seems that the auto-generated shared/js/ui/base/locale-resources.js
file fails the new Prettier linting checks. This file should probably live in
the build/ directory instead of the source tree, but for now let's at least
tweak the script that creates it to follow Prettier's rules.