asana_notify.json (#3155)
Adding a repository level list of Github usernames to assign/follow the pixel validation tasks in Asana. Eventually the plan is to also add any pixel owners with pixel errors to the follow list as well, but not yet. Those will not be added in the asana_notify.json but dynamically determined by the list of pixels with failures.


Co-authored-by: Jeanna Matthews <jmatthews@duckduckgo.com>